Wrote CO2Run.java
CO2Run.java is a class that holds a list of CO2 values.  It is an
abstract class intended to act as a parent for later defined CO2 value
holders.  These future classes are expected to include a MeanRun, which
will compute the average and standard deviation of a group of test
runs, and a FactorValuedRun, which may compute CO2 values over time
when given specific factor inputs.