Added classes FactorValuedRun.java and MeanRun.java
FactorValuedRun and MeanRun are subclasses of CalculatedRun and of
CO2Run.  FactorValuedRun represents a single simulation of GEOCARBSULF
for given factor values over an intrinsic time interval.  MeanRun takes
a list of CO2Runs of equal length and calculates an average run from
the average of all CO2 values at each time slice.